backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
meilisearch==0.29.0
python-dotenv==1.0.1
boto3==1.34.62
pandas==2.2.1
pyarrow==15.0.2
google-generativeai==0.3.1
duckdb==0.10.1
s3fs==2024.3.1
